[FIX] hr_appraisal: fix location of template field in department view
In a previous commit an invisible company_id field was added to the
department view. The xpath expression of appraisal_survey_template_id
relied on there only being one company_id field present in the
departemnt view and was therefore no longer correct after that change.

With this commit the xpath expression is updated to be more specific
as to which company_id field is used for reference.

task-3068250

closes odoo/enterprise#33948

X-original-commit: b35ca0ff9c78f2768ed4895ca38f4d2c9cbc281b
Signed-off-by: Kevin Baptiste <kba@odoo.com>