[FIX] account_winbooks_import: import analytic plans
The importing tool for Winbooks data was broken in 16, as this version
introduces Analytics Plans, which is also a mandatory field on
Analytic Accounts, and creating analytic accounts without plans would
throw an error.

The currency was also wrongly set to False, which was another cause for
error.

Some accounting automation flow was changed with the new version,
so this commit aims to adapt the import_move function.

The importing tool now works, and can also import analytic plans.

Winbooks file documentation:
https://help.winbooks.be/display/DEV/Annex+2++WinBooks+files+structure

opw-3032499

closes odoo/enterprise#33949

Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>