[FIX] web_cohort: Change text to avoid translation overwritten
This is a revert of a previous fix. In Odoo 16 the translation of python
code is not saved in the database. The translation is read directly from
the po files. Thus the previous workaround isn't needed anymore in Odoo
16.

Fix: undo the renaming of "Average", remove en.po and change it in pot
file

Related commit: https://github.com/odoo/enterprise/commit/fcfb4ff96647960e99f2b230b817f0c1a94c3e73

closes odoo/enterprise#34076

X-original-commit: cab1ed8a1d2d54e500bed92ddda92594d8dde501
Signed-off-by: Martin Trigaux (mat) <mat@odoo.com>