[IMP] account_sepa,account_sepa_direct_debit: add sepa mapping
Currently, it's impossible to have any non-latin characters in names and addresses when creating SEPA transactions. This can be a problem for many EU countries, such as Greece, Hungary, etc.

European Payments Council has published an official mapping for every non-latin character. With this commit, characters will be converted before generating an XML file.
Backport of odoo/enterprise#104735

was part of
task-2833976

closes odoo/enterprise#34019

X-original-commit: 20c60096ff9abe4ac32a15871b8e2d8a28d84e3b
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>