[FIX] account_reports: don't hide inactive records in groupby
When unfolding a groupby line, the groupby keys corresponding to inactive records were not displayed. For example, inactive customers did not appear in the Aged Receivable report.

This was due to the fact a call to .sorted() is done on the recordset containing the grouping keys, and this function calls .search() to fetch its results. A context key was hence missing in order to keep the archived records.

closes odoo/enterprise#33959

Signed-off-by: John Laterre (jol) <jol@odoo.com>