[IMP] stock_barcode*: instruction's message
Rewrites the BarcodeModel's getter `barcodeInfo` to have clearer code.
Previously, the BarcodeQuantModel didn't call the super anymore, but the
BarcodePickingModel did, so it doesn't make sense to keep a part of this
method in BarcodeModel.
Instead, moves the part from BarcodeModel into BarcodePickingModel, and
rewrites them to have something more common between the picking and the
quant models.

For the MainComponent `info` (the object containing instruction's data),
replaces the getter by a variable created in the template (thanks to
<t-set>) so that's way, the BarcodeModel method `barcodeInfo` is called
only one time instead of 6 times before (it was called every time the
getter was mentioned in the template, making useless extra computation
and being really annoying for debugging).

Also, removes the property `messageType` as it wasn't needed anymore and
cleans a little the template.

task-3048127

Part-of: odoo/enterprise#33600