[FIX] mrp_plm: No ECOs in BoM Overview for non-allowed users
If a user had at least "user" access to MRP but had no access to PLM
while it was installed, it would generate an error as soon as the user
would open the BoM Overview, as it would try to fetch data related to
the ECOs with a user that had no permissions to do it.

Now the display filter doesn't contain the ECOs if the user shouldn't
see them, and the data won't be fetched server-side for a non-allowed
user.

closes odoo/enterprise#34545

X-original-commit: 436e06a16e84a1561b7bb885c62b40bf21b2b1ee
Related: odoo/odoo#106935
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>