[FIX] account_accountant: Reconciliation wrong view
Before this PR, two flows going to the same view were broken. When "intrastat" is checked in the parameters:
* the first flow "Accounting -> Reconciliation -> (if the text "Good job ! ..." appears) Press on unreconciled entries -> not the desired view", the view displayed is not the one we want.
* The second flow "Chart of accounts -> configuration -> smart button Balance" the view displayed is also not the one we want.

In fact, before you check "intrastat" in the settings, a tree view is available thanks to the two flows explained above, but as soon as you check "intrastat", the tree view is replaced by another one because it will display the highest priority (the lowest number).

 Changing the priority of the intrastat tree view is enough to solve the problem. There is also no problem the other way around since the action is called directly when the particular tree view is needed.

 task-id: 3070967

closes odoo/enterprise#34037

X-original-commit: eb0d9149246bbce7df5be4eb30c26be234b01075
Signed-off-by: John Laterre (jol) <jol@odoo.com>