[IMP] sign: improve signature frame logic
This commit makes the signature frame be hidden when opening the signature
dialog in draw mode. The reasoning for this is that users try to fit into the
"lines" too much when signing, which adds friction as they will tend to try
several times. Therefore, we should let them use the canvas as they choose.

This commit also reduces the opacity of the frame, and adds a tooltip for the
frame checkbox, as the frame doesn't appear anymore when in draw mode.

task-3040855

X-original-commit: dfd512852ff34a951d8c0eb91b2e78a8b46fda8c
Part-of: odoo/enterprise#37338