[IMP] marketing_automation: handle tabs display out of kanban XML arch
This commit removes implementation details of the marketing_automation JS class
out of the XML arch to move all of the logic inside the JS code.

The reason for this is that the arch definition of views should be agnostic of
implementation details of its JS class.
This allows the view to work in cases where we don't spawn the JS class
(e.g: studio) and also makes it easier to extend / migrate.

Task-3070918

Part-of: odoo/enterprise#34049