[IMP] marketing_automation: store the bounce email and allow the user to read it
Purpose
=======
The bounce emails aren't stored in Odoo, which can complicated the
debugging of the email sending.

Now, we store this bounce email, and we allow the user to read it from
the interface, so he can easily find the issue when an email sending
fail.

Specifications
==============
The bounce email is stored of the mail notification for standard emails
sending, and on the mailing traces when using mass mailing.

Task-2116296
See odoo/odoo/pull/105923

closes odoo/enterprise#34051

Related: odoo/odoo#105923
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>