[IMP] sale_subscription: add first contract date field
before this commit, it was not possible to easily known when a contract first
started. Each renew has his own start_date but it is important to see easily
how old is a contract in reports.

There are several date fields on the sale order and we decided to create a new one
instead of reusing date_order. date_order has already two meaning depending on the
state of the order: create_date for quotations and order_date for confirmed orders.
We think that giving a third meaning to the same field would be confusing.

closes odoo/enterprise#34107

Taskid: 3071690
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>