[FIX] mrp_workorder: register component quantity
In some conditions, due to a server error, it is not possible to
validate a QC to register a consumed quantity

To reproduce the issue:
1. Create two storable products P_compo, P_finished
2. Set the on-hand quantity of both products to 1
3. Edit the operation type Manufacturing:
    - Reservation Method: Manually
4. Create a bill of materials BoM
    - Product: P_finished
    - Type: Manufacture
    - Operations: a new operation OP
    - Components:
        - P_finished, Consumed in OP
5. Create and confirm a MO with BoM
6. Open the tablet view of the work order
7. Set the quantities:
    - Producing: 1
    - P_compo's qty: 1
8. Validate the quality check

Error: an Odoo Error is displayed "AttributeError: 'quality.check'
object has no attribute 'product_uom_id'"

Before Odoo 15.5, the method `_create_extra_move_lines` was defined on
`mrp.workorder`, which has the field `product_uom_id`. From now on, it
is defined on the model `quality.check` which does not have such a
field.

OPW-3053568

closes odoo/enterprise#34045

Signed-off-by: Tiffany Chang <tic@odoo.com>