[FIX] sale_planning: prevent creation of invalid records in `action_plan_sale_order`
Prior to this fix, the creation of slot in `_add_slot_to_list` was performed without
validation.

It occurs that `_get_slots_values`, in some rare cases, tries to create a last slot
with having the `start_datetime` = `end_datetime`.

In order to prevent it, we added in check in `_add_slot_to_list`.

closes odoo/enterprise#34160

X-original-commit: 1a83d7e25cf1d244e32668d842a090f39562c098
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>