[FIX] account_report: cash flow comparison
The cash flow comparison was not working when there was a reconciled move in any of the periods.

The problem was due to a wrong initialization of two variables (reconciled_account_ids and reconciled_percentage_per_move). They should both have at least the column_group_key for each period, but they were initialize to empty values.

closes odoo/enterprise#35935

X-original-commit: 55011da14802d6e0ac73484f2d53a7ad88ee0bed
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>