[FIX] account_report: cash flow comparison
The cash flow comparison was not working when there was a reconciled move in any of the periods.

The problem was due to a wrong initialization of two variables (reconciled_account_ids and reconciled_percentage_per_move). They should both have at least the column_group_key for each period, but they were initialize to empty values.

closes odoo/enterprise#34053

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>