[FIX] test_l10n_be_hr_payroll_account: rectify test values
The test `test_student` does not work properly. The expected values are
based on the hourly wage of the contract without any rounding. However,
because `hourly_wage` is a monetary field, when writing any value in the
database, a rounding based on the related currency is applied on that
value (see `Monetary.convert_to_column`)

closes odoo/enterprise#34255

X-original-commit: abcbd78e8075d3267b7878faf29609b0bdfb48f0
Related: odoo/odoo#106278
Signed-off-by: Raphael Collet <rco@odoo.com>
Signed-off-by: Adrien Widart <awt@odoo.com>