[FIX] account_{batch_payment,sepa,sepa_direct_debit}: fix sepa mapping
Recently a sepa mapping was added in odoo/enterprise#33957.
Few characters were missing from the regex and they should be added for the complete mapping.
Characters <,>,& do not need to be in the mapping (and they're not) because they are correctly
formatted when an XML file is generated (&lt,&gt,&amp;)

closes odoo/enterprise#34159

X-original-commit: 3f55ad34339a6c9beaf9b993b6490aa5ec88fbf5
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>