[IMP] stock_barcode*: tour helper functions
Rewrites some helper's functions so they all use the same type of
parameters. Before this commit, those functions used sometime an jQuery
object, sometime an index (a number) or someting a `Node`.
Now, the given parameter can be either a `HTMLElement` (the line itself)
either an index (and it will find the corresponding line).

Renames `assertButtonIsVisible` into `assertButtonShouldBeVisible` and
removes `assertButtonIsNotVisible` as this method did nothing more than
calling `assertButtonIsVisible` with `shouldBeVisible` as false.

Renames `validateBarcodeForm` into `validateBarcodeOperation` as
`validateBarcodeForm` can be confusing as it's the actual picking who is
validated, from the client action which is not a view form.
Also, uses it everywhere it can be used.

Merges `assertLineQty` and `assertLineQuantityOnReservedQty` into the
same method since these methods were very close. Also, this method will
check the UoM if it's displayed.

Removes the `getLines` monkey patch in `stock_barcode_picking_batch` and
adds the part of getting lines by indexes in the `stock_barcode` helper.
Drops the `from`/`to` parameters as they were barely used and can easily
be replaced by the use of `index`.

Also, add some class or name attribute in the line's template to make it
easier and cleaner to find elements.

task-3048127

Part-of: odoo/enterprise#33600