[FIX] l10n_ch_reports: fix balance sheet's unaffected earnings
The line reporting the unaffected result from the previous years was not taking into account the amounts allocated directly from the current year earnings account, and only considered allocations directly made from the income/expense accounts. This was wrong, since we want to support both these cases.

Because of that, when doing the allocation this way, the result from previous years was reported twice: once in the "previous years' profit or loss" line, and once under "legal reserves", because of the account the allocation was made to.

In the meantime, we improve the name of this line, so that it betters reflect what it actually computes.

OPW 2731621

closes odoo/enterprise#34413

X-original-commit: c403b946929959e258e872aaa479b5e4c98ac008
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>