[FIX] account_asset: make report more robust
The aim of this commit is to allow the customer to delete a column of the asset
report without it to crash.

context:
A customer deletes one column of the asset report because he doesn't need it
and then try to open it.

Previous to this commit:
The report crashes because of an index error.

After this commit:
The report works correctly but doesn't display the deleted column

closes odoo/enterprise#34142

Task: 3072503
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>