[FIX] web_studio: grouped kanban editor with sample in arch
Have a kanban arch with sample="1" set on the kanban node of the arch.
The kanban must be grouped by a many2one field. Also, no data should be retrieved
hence activating sample data.

Open that kanban in studio.

Before this commit, there was a crash, because sample data are not meant to be dealt with
by studio. Instead, when no data is present a fake empty record should be shown.

After this commit, there is no crash.

closes odoo/enterprise#34140

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>