[IMP] sale_subscription: auto close unpaid subscriptions
Before this commit, when a subscription did not had any payment method (token),
the invoice would be posted and handled by the accounting team. Confirmed
invoice could take some time to be reconcilied and long discussion could happen
between the customer and the company. During this time, the subscription would
remain in progress and the service would be provided. This commit ensures that
the subscription is closed if the invoice is not paid in time.

The maximum delay for payment is the latest date_maturity entry on the
account.move.line of the invoice + auto_close_limit in days saved on the
quotation template (default is 15 days). After this delay, the contract is
closed.

Taskid: 3054994
Part-of: odoo/enterprise#35710