[IMP] sign: improve draw mode in signature dialog
Before this commit, if the signature item was already filled with a signature,
the signature dialog would be opened in draw mode and would load the signature.
However, if you're clicking on the signature item, you are probably trying to
change the signature. Therefore, it was decided to keep opening it in draw
mode, but with a clear canvas. In case the signature item is not filled, the
dialog will keep being opened in auto mode.

task-3040855

closes odoo/enterprise#37338

X-original-commit: 50b35c6961b774d0f358c5c5c17d279f315a28c8
Related: odoo/odoo#113248
Signed-off-by: Arnaud Joset <arj@odoo.com>
Signed-off-by: Leonardo Pavan Rocha <lpr@odoo.com>