[FIX] mr_workorder_plm: Keep suggested title for new step suggestion
When suggesting a new step in the tablet view, a user can write a title,
a set of instructions and add a document / image.
But when PLM is installed and the changes are also added to an ECO, the
content of the title in the suggested step is replaced by "New step from
manufacturing feedback".

This change aims to keep the suggested title while still notifying that
this was a step suggested in the tablet view.

Part of task-2985735

closes odoo/enterprise#34212

X-original-commit: fb1a22bf3f1610ec12285baac74b07f4cf532466
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>