[FIX] knowledge: emoji access crash on readonly
This commit fixes a crash of the application when trying to open an article
on which the user can only read, or when trying to open another article from
one that he can only read.

This was linked to the fact the new openArticle function tried to access the
emoji button of the said article even though it doesn't exist for the user
that is only able to read.

Now we check if the button exists before trying to modify it.

task-3076075

closes odoo/enterprise#34275

Signed-off-by: David Beguin (dbe) <dbe@odoo.com>