[FIX] documents_spreadsheet: fix spreadsheet creation form template
This commit addresses two bugs:

1. A call to `ormService` when trying to save a spreadsheet as
template was still using the old API and would crash when trying to
convert a pivot with several levels of group by.

2. The conversion of a template to a standard spreadsheet would
traceback when it contained a pivot without user data; i.e. a pivot
for which the combination of model/domain would not have any data.

opw 3060534

closes odoo/enterprise#34194

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>
Co-authored-by: Lucas Lefèvre <lul@odoo.com>