[FIX] l10n_cl_edi_stock: changed test dates to match changes on XML
Also fixed test which wasn't generating move lines by adding quantity_done on pickings
We now monkey patch _get_cl_current_datetime instead of _get_cl_current_strftime since when we need to convert from Datetime to Date we call the former directly.
This forces us to use a timezone aware Datetime because comparisons between Datetimes are made in other parts of the code and you cannot compare a timezone aware Datetime with a timezon naive one
We use 'America/Santiago' with pytz.timezone.localize so that the results of the tests won't change depending on the time of the year they are ran.
We freeze time more strictly now, in case another process interferes with the tests in the future

closes odoo/enterprise#34590

X-original-commit: 81572b5bcf35503da63cf97567f3e3602b936105
Signed-off-by: William André (wan) <wan@odoo.com>