[FIX] social(_*): add test before formatPost call
When using studio on the social.stream.post kanban view it
crashes because formatPost doesn't exist in the context (studio
doesn't use the js_class). So we add a test to check if the
method exist before calling it.
This will be changed for v16.1 by a custom field widget.

Task-3061215

Part-of: odoo/enterprise#33806