[FIX] social(_*): add test before formatPost call
When using studio on the social.stream.post kanban view it
crashes because formatPost doesn't exist in the context (studio
doesn't use the js_class). So we add a test to check if the
method exist before calling it.
This will be changed for v16.1 by a custom field widget.

Task-3061215

X-original-commit: 7f639112f9c743f927c12da99a25f58e2e84622a
Part-of: odoo/enterprise#34687