[FIX] web_studio: kanbanEditor prevent click on anything other than an edition feature
In studio in general, we don't want clicks to honor their default behavior (open records etc...)
Instead, clicks are usually a ways to access either a node's properties or an edition feature
(add field, structures etc...).

Before this commit, in kanban, when the arch contained `t-on-click` (which it should absolutely not btw),
and that one clicked on such a node in the DOM, a crash occured.

After this commit, every click is captured and cancelled if it doesn't concern an edition feature.

closes odoo/enterprise#34198

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>