[FIX] web_studio: buttonBox button hook at the right place
Have a form arch with a buttonBox (`<div class="oe_button_box" />`) already present. To make the bug more visually salient, the first button of the buttonBox should be invisible.

Before this commit, the hook to add a button to the box was invisible. This was because that hook
was placed inside the first slot, after the first button.

After this commit, it is placed in the first slot of the box and should be rendered anyway.

opw-3076683

closes odoo/enterprise#34202

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>