[FIX] account_reports: fix totals below sections with account groups
Steps to reproduce:
- Install Belgian accounting.
- Activate totals below sections in the settings.
- Create and confirm an asset in the previous year.
- Go to the depreciation schedule.

Observed behavior:
The asset line is displayed after all the account group lines
AND the totals of those lines.

Desired behavior:
The asset line should be displayed after the group lines and
BEFORE the totals of those lines.

If you add other assets with depreciation accounts having
a common root, the entire hierarchy will be duplicated for
every asset. Even the groups that are shared in the hierarchy.

task-3041037

closes odoo/enterprise#33294

Related: odoo/odoo#105788
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Co-authored-by: Olivier Colson (oco) <oco@odoo.com>