[FIX] mrp_workorder: always send an object to onClose()
The onClose() param of doAction is bind in tablet.js to rerender
everything after closing a modal. In some cases, a notification is
displayed thanks to a message passed throught this bind method. The
issue was that the notification message was given as the only plain
string argument of the onClose method while it sometimes receives an
object (from the Discrad button for instance).

This commit ensure the bind method always receives an object. If the
'message' key is in, it display the value in a notification.

closes odoo/enterprise#34261

X-original-commit: 2099e8e4bd2032a06fb3d354a117b49973f33c24
Signed-off-by: William Henrotin (whe) <whe@odoo.com>