[FIX] project_timesheet_forecast: generate the right number of timesheets from planning
Steps to reproduce:
- Create a product as a service, invoicing policy based on timesheet, project and task created on order
- Create a sale order with 5 hours of the service, and confirm it
- Select a shift in planning, it will create two slots with respectively 1 and 4 hours allocated.
- Publish the slots created and generate timesheets

Issue:
The generate action creates two timesheets with 5 hours each (in sum 10 hours) but we expect only one
with 5 hours.

Cause:
The action manage to create the minimum number of timesheets based on the date of the planning slot.
But it doesn't check if the date of a slot from and employee has already been added to the data used to
create timesheets.

Solution:
Increment the count of timesheet per date to ensure we have one timesheet per date per employee to create

opw-3070859

closes odoo/enterprise#34230

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>