[FIX] account_3way_match: do not compute release_to_pay for non-invoices
This field is not necessary on non invoice related journal entries.

A side effect of this, and the reason why it was detected was that
setting `release_to_pay_manual` in `_compute_release_to_pay` triggers a
`write` because the compute method is not supposed to modify that field.
The `write` then triggered the computation of other fields, including
the `name`. In the case of the PoS[1], entries could not be balanced while
in a transient state (in memory) but flushing the record would raise an
error.

[1]: see test_correct_partner_on_invoice_receivables
[1]: see https://runbot.odoo.com/runbot/build/21236209

closes odoo/enterprise#34297

X-original-commit: 125f1c7bd4269fad24b853bb95ccb6688ba1402c
Related: odoo/odoo#106325
Signed-off-by: Quentin De Paoli <qdp@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>