[FIX] web_studio: create inline view if current arch automatically inlined by server
The new get_views API inlines every subview into the main view in an effort to
optimize client/server communication.

In the situation:
```
<form>
  <field name="my_one_to_many" />
</form>
```

The server algorithm will fetch kanban or list views to display that x2many and inline them
in the final arch that is sent to the client.

Before this commit though, Studio was unable to determine whether that inlined subview
has been put there by the server or was present in the original arch.

Namely when entering the list view editor for that one2many and editing it, there was a crash saying, justifiably
that the tree view did not exist.

After this commit, we signal to the client that the subview was inlined and that consequently we must
create an inline view in order to edit it.

X-original-commit: 0f6a6e9654229c202550938ef0c2a0bf9ed01b0e
Part-of: odoo/enterprise#34376