[FIX] mrp: prevent editing of readonly locations for MO moves
A MO.move_raw_ids.location_id should never be directly set by users in
the form view. Now that this field is either readonly when multi-loc is
active or not in the view at all when multi-loc is not active, we remove
the setting of this field from tests that fail with this fix.

COM PR: odoo/odoo#105959

closes odoo/enterprise#34396

X-original-commit: c57e1272b66df77e6528c7bf1ba800fe67080012
Related: odoo/odoo#106542
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Tiffany Chang <tic@odoo.com>