[FIX] sale_subscription: keep draft invoice after tx callback succeeded
When invoicing subscription with a payment token in automatic mode
(i.e. cron), once the invoice has been created we're going to:

1. issue a payment request (`_do_payment()`) and commit

   This will call the transaction callback method, in our case
   `reconcile_pending_transaction()`, which will update the next
   invoice date (1st time)

2. handle the result of that request (either success or failure) and
   commit

3. Later, the transaction post-processing will see the transaction
   and process it.

In case the payment request succeed (1), but the method that handle it's
result crash (2) we're going to unlink the draft invoice.

As we've committed after the payment succeeded but unlinked the invoice,
when the transaction the transaction is post-processed (3) it won't have
any invoice; so we're creating a new invoice and forcing update of the
next invoice date (2nd time).

The result is that the subscription next invoice date is updated twice.

This commit ensure we don't delete the invoice if the payment succeeded
and the callback method as been called, even if later down the road the
crash for some other reason.

  The transaction post-processing will then see a transaction with a
"draft" invoice, that he's going to post before creation the payment
and reconciling them; but will NOT update the next invoice date.

opw-3051424

closes odoo/enterprise#34310

X-original-commit: b80b71562f57fbcc1fdf40ecb5d132b3a20a5bad
Signed-off-by: Arnaud Joset <arj@odoo.com>
Signed-off-by: Xavier Alt (xal) <xal@odoo.com>