[FIX] account_reports: fix hierarchy filter display
Before this commit, if you open the balance sheet
of a company that has no account groups,
the "Hierarchy and Subtotals" option will be available,
but it will have no effect (how could it).

But if was confusing to see an option and see no effect when
clicking on it.

So the option will only be displayed if the company
has at least one account group.

Part-of: odoo/enterprise#33294