[FIX] account_reports: apply proper date computation in options if no mode is provided + clean dead code
Providing dates without the mode in the previous_options lead to the default_filter always being applied. This was wrong and could cause issues when calling the report directly from the code (for example to generate an export file).

Also, removed an unused condition containing dead code (self.filter_date doesn't exist anymore). When using a 'custom' date filter, the dates will be computed before reaching this condition anyway, so nothing is to be done there.

X-original-commit: 6e29e42ef16162bcfa1c54e8a16f57329d2d689f
Part-of: odoo/enterprise#34366