[FIX] l10n_be_reports: use proper dates when exporting tax report XML + fix file indentation
The date options was not passed properly to the report. Because of that, though the file contained the right date, the data it contained where always the ones from the report's default date filter (last_month).

Furthermore, he indentation of the generated file was messy. This commits brings it back to something structured and human-readable.

OPW 3073193

closes odoo/enterprise#34366

X-original-commit: 15b37cc20059454a2e5ad5d27b344941d08b9561
Signed-off-by: Laurent Smet <las@odoo.com>