[FIX] l10n_cl_edi: Enhance currency calculations and precision in the DTE rendering
This commit improves the calculations when we have tax included in price.
It also include more precision in product unit prices. Also we can create
invoices using a different currency. Odoo will calculate the value based
on the account entry for CLP and still preserve the amount in other currencies.
It also fixes the precision for export invoices.
From now on, the MntBruto tag is not used in the DTEs as we pass everything in net amounts.

X-original-commit: 4640ea3a0124e21d78406ed031c19990d354d065

In comparison with the original commit, we use amount_currency instead of
price_subtotal as the price_subtotal is not set anymore on tax lines.
Also the tests needed adaptations as they were moved.

closes odoo/enterprise#46756

X-original-commit: aa441bf101b961078c91a207c20de3c09a0b3e42
Signed-off-by: Josse Colpaert <jco@odoo.com>