[IMP] sale_subscription: improve behavior of free subscription
Before this commit, when a subscription was free, it would not be invoiced
by the invoicing cron. A subscription was condifered free if the sum of
the line subtotal was equal to 0. It was not sufficient as some orders
contains both recurring and non recurring lines.

This PR improve the behavior by displaying different behaviors for recurring
and non recurring lines.

When the sum of recurring line is 0, the subscription is free.
When a non recurring line balance the reucrring one, the subscription
is invoiced but the amount of the invoice is equal to 0.

Taskid: 3082618
Part-of: odoo/enterprise#35100