[FIX] website_appointment: match navbar and chevrons color.
To reproduce:

Install artist theme on website, go to appointment submenu.
The end of chevrons is a different color than the rest of
the navbar. This is because the navbar color is different
than the background color, used in the colors of the navbar
of appointment.

Fix:

Remove the color aspect of the navbar. Therefore, the
appointment navbar is now independant of the one above.
It has coherent chevrons, and is more readable than before,
especially for dark navbars.

Task-3080093

closes odoo/enterprise#34842

X-original-commit: b0796e50a6a6dc65fb067ffe8777bf7decb2eeb0
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>