[FIX] l10n_de_skr03: translation
Ideally, we develop in English any localization (then translate to the domestic language, in this case German) All the module has been translated in english in this PR: https://github.com/odoo/enterprise/pull/34313

By translating the skr03 module this test was broken. Since it search on the name in german that is now in english.

closes odoo/enterprise#34323

Task-id: 3059115
Related: odoo/odoo#106324
Signed-off-by: Laurent Smet <las@odoo.com>