[FIX] web_studio: edit one2many list with records and fieldDependencies
Have a form view with an embedded list view.
The list view should have some records and a field having widget, itself having fieldDepencies
(that is, fields of the same model on which the widget depends).

Before this commit, there were crashes beacause the definition of the dependency field was not updated
to the one declared python side.
More generally, the fields get of the related model was never taken into account when editing the X2Many.

After this commit, the fields of the related model are updated with the full list of fields declared server side.
And there are no crashes anymore.

closes odoo/enterprise#34250

Related: odoo/odoo#106271
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>