[FIX] stock_barcode: scan package in dest
How to reproduce:
- Enable package and multilocation;
- Create a package;
- For a product, add some quantity on hand in the created package in the
  default location (WH/Stock usually);
- Open the Barcode App;
- At the main menu, scan the location's barcode to create an internal
  picking in this location ("WH-STOCK" for WH/Stock);
- Scan the package -> It doesn't nothing but display a notification
  saying the user has to scan a product or a package in the current
  location.

Expected behavior:
As the scanned package is in the right location, it should create a
barcode line for each quant of the scanned package (or a package line if
the picking type uses "Move Entire Packages").

The issue is, as the package is also in the picking's dest location, it
tries to use the scanned package as the result package (and so want to
apply it on previous scanned line if there is some) instead of use its
quants to create barcode lines.

closes odoo/enterprise#34718

X-original-commit: 6f4850e00af16a397232d533d50d8d398a849f3c
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>