[FIX] account_accountant: Fix partner_id set on statement line on validate
- Open the bank reconciliation widget
- Select a statement line with a partner_id set
- Click on the validate button without changing anything
=> the partner is gone

Since the 'auto_balance' line hasn't be edited and the wizard has been validated directly, its flag is still 'auto_balance'.
During the validating, we are getting the partners to set on the statement line but the 'liquidity' line and the 'auto_balance' line are ignored.
Therefore, we don't find any partner to set on the statement line and the original partner is now gone.

closes odoo/enterprise#34340

Issue: 3075557
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>