[FIX] stock_barcode: scan picking name in kanban
In the Barcode App, when a picking's type operation is selected, it
displays the picking's kanban view. On the top of that, there is a
message, "Scan a transfer or a product to filter your records".
This message made sense when the focus was by default on the searchbar,
but since #31306, it's not the case anymore as we prefer to catch the
scan event to call `filter_on_barcode`.
This method, `filter_on_barcode`, filters the pickings by the scanned
product (also by the package but it isn't written in the message) but
doesn't filter by the picking's name.
This commit fixes that, so the user can scan the picking's barcode to
display only this picking.

task-3081480

closes odoo/enterprise#34377

Signed-off-by: William Henrotin (whe) <whe@odoo.com>